fix: rename misc package to correct name
