--- misc/src/main/java/love/broccolai/corn/core/package-info.java
@@ -1,4 +1,0 @@
-/**
- * General utilities
- */
-package love.broccolai.corn.core;
