+++ misc/src/main/java/love/broccolai/corn/misc/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * General utilities
+ */
+package love.broccolai.corn.misc;
