--- misc/src/test/java/love/broccolai/corn/core/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Core tests
- */
-package love.broccolai.corn.core;
