+++ misc/src/test/java/love/broccolai/corn/misc/NumbersTest.java
@@ -1,4 +1,4 @@
-package love.broccolai.corn.core;
+package love.broccolai.corn.misc;
 
 import org.junit.jupiter.api.Test;
