+++ misc/src/test/java/love/broccolai/corn/misc/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Core tests
+ */
+package love.broccolai.corn.misc;
